# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Course",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order/create": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "create a new order",
                "description": "Create a new order",
                "operationId": "create_order",
                "parameters": [
                    {
                        "name": "payment_method",
                        "in": "query",
                        "description": "Payment method: banking, vnpay",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "voucher_id",
                        "in": "query",
                        "description": "Voucher ID to apply to order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/order/confirmpayment/{order_id}/{order_code}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Check your status or change",
                "description": "Check your status or change",
                "operationId": "confirmpayment1",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/order/checkpayment/{order_id}/{order_code}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "checkpayment",
                "description": "checkpayment",
                "operationId": "checkpayment",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/order/{order_code}/check_amount": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "check_amount",
                "description": "check_amount",
                "operationId": "check_amount",
                "parameters": [
                    {
                        "name": "order_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/order/my_order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "myorder",
                "description": "myorder",
                "operationId": "myorder",
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}